pyproject.toml
[project]
name = "de-with-python-uv"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apache-airflow-providers-google==10.24.0",
    "apache-airflow>=2.10.4",
    "pandas>=2.1.2,<=2.1.4",
    "psycopg2-binary>=2.9.10",
    "google-cloud-storage>=2.19.0",
]
